pyproject.toml
[project]
name = "ads-gen"
version = "0.1.0"
description = "AI-powered advertisement video generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "rembg>=2.0.0",
    "onnxruntime>=1.23.2",
]

[project.scripts]
ads-gen = "backend.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
